Refactor file paths to adhere to core structure
