+++ src/core/options-merger.ts
